Extracted serde code for Dock into separate file
